IBX-538: Renamed EzSystemsPlatformInstaller Bundle & Extension to IbexaRepositoryInstaller
