+++ src/bundle/Installer/IbexaRepositoryInstallerBundle.php
